fix: 타ì입 변경

+++ backend/src/main/java/org/scoula/finalreport/dto/FinalReportDTO.java
@@ -14,8 +14,8 @@ public class FinalReportDTO {
     private String jeonseRatioRating;
     private Double jeonseRatio; // 전세가율
     private Double regionAvgJeonseRatio; // 지역 평균 전세가율
-    private Long deposit; // 보증금
-    private Long expectedSellingPrice; // 예상 매매가
+    private Integer deposit; // 보증금
+    private Integer expectedSellingPrice; // 예상 매매가
     private List<Boolean> checked; // 체크 항목 상태
     private String checklistRating;
 }
