+++ backend/src/main/java/org/scoula/finalreport/dto/FinalReportRawDTO.java
@@ -10,8 +10,8 @@ public class FinalReportRawDTO {
     private String jeonseRatioRating;
     private Double jeonseRatio;
     private Double regionAvgJeonseRatio;
-    private Long deposit;
-    private Long expectedSellingPrice; // 예상 매매가
+    private Integer deposit;
+    private Integer expectedSellingPrice; // 예상 매매가
     private String checked; // json 문자열 그대로
     private String checklistRating;
 }
